implement modeldb-sqlite with the shared class
