+++ packages/modeldb-sqlite-shared/src/index.ts
@@ -1,0 +1,3 @@
+export { AbstractSqliteDB, Query, Method } from "./AbstractSqliteDB.js"
+export { SqlValue } from "./encoding.js"
+export { ModelAPI } from "./ModelAPI.js"
