+++ packages/modeldb-sqlite/package.json
@@ -17,6 +17,7 @@
 	},
 	"dependencies": {
 		"@canvas-js/modeldb": "0.11.0",
+		"@canvas-js/modeldb-sqlite-shared": "0.11.0",
 		"@canvas-js/utils": "0.11.0",
 		"@ipld/dag-json": "^10.2.2",
 		"@types/better-sqlite3": "^7.6.10",
